@@ -1,4 +1,4 @@
 import React from 'react';
-import { Time } from './Time';
+import { Time } from './Time.js';
 
 export const App = () => <Time />;
